requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.100.0
fastapi-cli==0.0.4
filelock==3.14.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==4.1.0
mypy-extensions==1.0.0
nodeenv==1.9.0
orjson==3.10.3
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pottery==3.0.0
pre-commit==3.7.1
pydantic==1.10.15
pydantic_core==2.18.3
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==4.6.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlmodel==0.0.18
starlette==0.27.0
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
ulid==1.1
ulid-py==1.1.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
